JavaScript Contact Form with Character Counter
JavaScript Contact Form with Character Counter

This project is a Contact Form, which is built using JavaScript. A special feature of this form is that it provides a Character Counter, which shows the number of characters remaining to the user while typing in a particular input field. This helps improve the user experience of the form.